refactored listener to its own class so that it can be called from other exts
